pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hapr"
version = "0.1.0"
description = "HAProxy Audit & Reporting Tool — security baseline scoring, TLS scanning, CVE checking, and interactive reporting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "HAPR Contributors"},
]
keywords = ["haproxy", "security", "audit", "tls", "cve", "hardening"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Networking",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "plotly>=5.18",
    "jinja2>=3.1",
    "rich>=13.0",
    "sslyze>=6.0",
    "nvdlib>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
hapr = "hapr.cli:cli"

[project.urls]
Homepage = "https://github.com/hapr-tool/hapr"
Issues = "https://github.com/hapr-tool/hapr/issues"

[tool.setuptools.packages.find]
include = ["hapr*"]

[tool.setuptools.package-data]
hapr = ["../framework/*.yaml", "../templates/*.j2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
